Renamed PriceCommands plugin to ChatCommands
