+++ runelite-client/src/main/java/net/runelite/client/plugins/PluginManager.java
@@ -36,6 +36,7 @@ import net.runelite.client.RuneLite;
 import net.runelite.client.plugins.account.AccountPlugin;
 import net.runelite.client.plugins.boosts.Boosts;
 import net.runelite.client.plugins.bosstimer.BossTimers;
+import net.runelite.client.plugins.chatcommands.ChatCommands;
 import net.runelite.client.plugins.clanchat.ClanChat;
 import net.runelite.client.plugins.cluescrolls.ClueScrollPlugin;
 import net.runelite.client.plugins.combatnotifier.CombatNotifier;
@@ -52,7 +53,7 @@ import net.runelite.client.plugins.jewelrycount.JewelryCount;
 import net.runelite.client.plugins.mousehighlight.MouseHighlight;
 import net.runelite.client.plugins.opponentinfo.OpponentInfo;
 import net.runelite.client.plugins.pestcontrol.PestControl;
-import net.runelite.client.plugins.pricecommands.PriceCommands;
+import net.runelite.client.plugins.chatcommands.ChatCommands;
 import net.runelite.client.plugins.rememberusername.RememberUsername;
 import net.runelite.client.plugins.runecraft.Runecraft;
 import net.runelite.client.plugins.woodcutting.WoodcuttingPlugin;
@@ -104,7 +105,7 @@ public class PluginManager
 		plugins.add(new FishingPlugin());
 		plugins.add(new WoodcuttingPlugin());
 		plugins.add(new RememberUsername());
-		plugins.add(new PriceCommands());
+		plugins.add(new ChatCommands());
 		plugins.add(new ClueScrollPlugin());
 
 		if (RuneLite.getOptions().has("developer-mode"))
